+++ include/nfx/Silicon.h
@@ -47,6 +47,7 @@
 
 // 74xx packages
 #include "nfx/silicon/gate/sn74xxx/LS74.h"
+#include "nfx/silicon/gate/sn74xxx/LS373.h"
 
 // Signal layer
 #include "nfx/silicon/signal/Bus.h"